Working

+++ app/api/comments_routes.py
@@ -2,8 +2,20 @@ from flask import Blueprint, request
 from flask_login import login_required, current_user
 from app.models.comment import Comment
 from app.models import db, User
+from app.forms import CommentForm
 
 comments_routes = Blueprint('comments', __name__)
+
+def validation_errors_to_error_messages(validation_errors):
+    """
+    Simple function that turns the WTForms validation errors into a simple list
+    """
+    errorMessages = []
+    for field in validation_errors:
+        for error in validation_errors[field]:
+            errorMessages.append(f'{field} : {error}')
+    return errorMessages
+
 
 @comments_routes.route('/<int:id>', methods=['PUT'])
 @login_required
@@ -11,23 +23,31 @@ def fix_comment(id):
     """
     Query for a single comment and edit that comment
     """
-    data = request.json()
     comment = Comment.query.get(id)
-    comment.body = data['body']
-    db.session.add(comment)
-    db.session.commit()
-    return comment.to_dict()
+    if current_user.id == comment.user_id:
+        form = CommentForm()
+        form['csrf_token'].data = request.cookies['csrf_token']
+        print(form.data)
+        if form.validate_on_submit():
+            comment.body = form.data['body']
+            db.session.add(comment)
+            db.session.commit()
+            return comment.to_dict()
+        return {'errors': validation_errors_to_error_messages(form.errors)}, 401
+    return {'errors': ['Unauthorized']}
 
 @comments_routes.route('/<int:id>', methods=['DELETE'])
 @login_required
-def remove_comment():
+def remove_comment(id):
     """
     Query for all comments for a story and returns them in a list of dictionaries
     """
     comment = Comment.query.get(id)
-    db.session.delete(comment)
-    db.session.commit()
-    return "Successfully Deleted"
+    if current_user.id == comment.user_id:
+        db.session.delete(comment)
+        db.session.commit()
+        return comment.to_dict()
+    return {'errors': ['Unauthorized']}
 
 # ====================likes comments====================================
 
@@ -43,9 +63,9 @@ def get_like(id):
         'num':num,
         'allUser':[(user.id) for user in all_liked_user]
     }
-    
+
     return num_like
-    
+
 #post like comment
 @comments_routes.route('/<int:id>/likes', methods=['POST'])
 @login_required
@@ -53,7 +73,7 @@ def post_like(id):
     comment = Comment.query.get(id)
     like_comment_user = User.query.get(current_user.id)
     all_liked_user = comment.liked_comment_user.all()
-    
+
     if not all_liked_user:
         comment.liked_comment_user.append(like_comment_user)
     else:
@@ -62,7 +82,7 @@ def post_like(id):
                 return "You already left a comment on this story"
             else:
                 comment.liked_comment_user.append(like_comment_user)
-                
+
 
     # the number of like for the comment
     num = comment.liked_comment_user.count()
@@ -73,7 +93,7 @@ def post_like(id):
         'story_id': comment.story_id,
         'num':num
     }
-      
+
     return num_like
 
 #delete like of one comment,  for the unlike button
